[Statie] move Guard test to propper test case
