+++ packages/Statie/packages/Generator/tests/Configuration/GeneratorElementGuardSource/InvalidLectureSorter.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\Statie\Generator\Tests\Configuration\GeneratorElementGuardSource;
+
+final class InvalidLectureSorter
+{
+}
