--- packages/Statie/packages/Generator/tests/ObjectSorter/ObjectSorterSource/InvalidLectureSorter.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\Statie\Generator\Tests\ObjectSorter\ObjectSorterSource;
-
-final class InvalidLectureSorter
-{
-}
